+++ packages/react/src/components/Flex/FlexItem.tsx
@@ -7,7 +7,10 @@ import { mergeStyles } from '../../lib/mergeThemes'
 import { ComponentSlotStylesPrepared } from '../../themes/types'
 
 type ChildrenFunction = (
-  params: { styles: ComponentSlotStylesPrepared; classes: string },
+  params: {
+    styles: ComponentSlotStylesPrepared
+    classes: string
+  },
 ) => React.ReactElement<any>
 
 export type FlexItemChildren = React.ReactElement<any> | ChildrenFunction