+++ packages/react/src/lib/accessibility/Behaviors/Chat/chatMessageBehavior.ts
@@ -8,6 +8,7 @@ import { FocusZoneTabbableElements, FocusZoneDirection } from '../../FocusZone'
  * Sets the message to be a focusable element.
  * Adds a vertical circular focus zone navigation where a user navigates using a Tab key.
  * Adds a key action which prevents up and down arrow keys from navigating in FocusZone, we only want a Tab key to navigate.
+ * Adds an escape key action which focuses the chat message, i.e., moves key handling from inside a message back to the chat list.
  *
  * @specification
  * Embeds component into FocusZone.
@@ -34,6 +35,9 @@ const chatMessageBehavior: Accessibility = () => ({
       // prevents default FocusZone behavior, in this case, prevents using arrow keys as navigation (we only want a Tab key to navigate)
       preventDefault: {
         keyCombinations: [{ keyCode: keyboardKey.ArrowUp }, { keyCode: keyboardKey.ArrowDown }],
+      },
+      focus: {
+        keyCombinations: [{ keyCode: keyboardKey.Escape }],
       },
     },
   },